Add files via upload
In this code, a classifier of handwriting images based on MNIST is built. The data is taken from the Kaggle website (https://www.kaggle.com/competitions/digit-recognizer/data ). The gradient histogram (HOG) feature was used. The quality criterion of the classifier is the accuracy metric. In the future, it is necessary to achieve that the accuracy metric is equal to 1.